factors out test_dezrann.py into folder dedicated to tests on Mozart
